requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-chroma>=0.0.1
chromadb==0.5.23
tokenizers>=0.13.2,<=0.20.3
transformers==4.31.0
zhipuai>=2.0.1
python-dotenv